Add food crud

+++ app/Http/Controllers/FoodController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Food;
 use Illuminate\Http\Request;
+use App\Inventory;
 
 class FoodController extends Controller
 {
@@ -12,9 +13,10 @@ class FoodController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Inventory $inventory)
     {
-        //
+        $this->authorize('update', $inventory);
+        return view('private.food.create', compact('inventory'));
     }
 
     /**
@@ -25,7 +27,27 @@ class FoodController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $this->validate($request, [
+            'inventory_id' => 'required|integer|exists:inventories,id',
+            'name' => 'required|string',
+            'expiration_date' => 'required|date'
+        ]);
+
+        $inventory = Inventory::find($request->input('inventory_id'));
+
+        $this->authorize('update', $inventory);
+
+        $food = new Food();
+        $food->inventory_id = $inventory->id;
+        $food->name = $request->input('name');
+        $food->expiration_date = $request->input('expiration_date');
+        $food->save();
+
+        return redirect(route('inventories.show', $inventory))
+                ->with('alert', [
+                    'type' => 'success',
+                    'content' => "{$food->name} was added correctly to {$inventory->name} inventory"
+                ]);
     }
 
     /**
@@ -36,6 +58,7 @@ class FoodController extends Controller
      */
     public function show(Food $food)
     {
+        $this->authorize('view', $food->inventory);
         return view('private.food.show', compact('food'));
     }
 
@@ -47,7 +70,9 @@ class FoodController extends Controller
      */
     public function edit(Food $food)
     {
-        //
+        $inventory = $food->inventory;
+        $this->authorize('update', $inventory);
+        return view('private.food.edit', compact('food', 'inventory'));
     }
 
     /**
@@ -59,7 +84,26 @@ class FoodController extends Controller
      */
     public function update(Request $request, Food $food)
     {
-        //
+        $this->validate($request, [
+            'inventory_id' => 'required|integer|exists:inventories,id',
+            'name' => 'required|string',
+            'expiration_date' => 'required|date'
+        ]);
+
+        $inventory = Inventory::find($request->input('inventory_id'));
+
+        $this->authorize('update', $inventory);
+
+        $food->inventory_id = $inventory->id;
+        $food->name = $request->input('name');
+        $food->expiration_date = $request->input('expiration_date');
+        $food->update();
+
+        return redirect(route('inventories.show', $inventory))
+            ->with('alert', [
+                'type' => 'success',
+                'content' => "{$food->name} was updated correctly on {$inventory->name} inventory"
+                ]);
     }
 
     /**
@@ -70,6 +114,7 @@ class FoodController extends Controller
      */
     public function destroy(Food $food)
     {
+        $this->authorize('update', $food->inventory);
         $food->delete();
 
         return redirect(route('inventories.show', $food->inventory))
